Move babel plugins.
