+++ packages/core/src/babelPlugins/__tests__/__fixtures__/cell/cell-with-default-and-other-named-export/code.js
