+++ packages/core/src/babelPlugins/__tests__/__fixtures__/cell/cell-with-default-export/code.js
