+++ packages/core/src/babelPlugins/__tests__/__fixtures__/import-dir/__fixtures__/a.js
