+++ packages/core/src/babelPlugins/__tests__/__fixtures__/import-dir/import-dir-default/code.js
