+++ packages/core/src/babelPlugins/__tests__/__fixtures__/routes-auto-loader/routes-with-no-imports/output.js
