+++ packages/core/src/babelPlugins/__tests__/babel-plugin-redwood-routes-auto-loader.test.ts
