+++ packages/core/src/babelPlugins/babel-plugin-redwood-routes-auto-loader.ts
